course modification

+++ feign-api/src/main/java/edu/codebridge/feign/code/IdentityCode.java
@@ -4,8 +4,8 @@ package edu.codebridge.feign.code;
  * This class is the const to denote the type of user(teacher student admin etc.)
  */
 public class IdentityCode {
-    public static final Integer STUDENT = 0;
-    public static final Integer TEACHER = 1;
-    public static final Integer ADMINISTRATOR = 2;
-    public static final Integer UNAUTHORIZED_TEACHER = 3;
+    public static final Short STUDENT = 0;
+    public static final Short TEACHER = 1;
+    public static final Short ADMINISTRATOR = 2;
+    public static final Short UNAUTHORIZED_TEACHER = 3;
 }
